#3 Inherited monsters are also monsters
The Skip_Super attribute and Super:: reference is unhandled. Therefore: a) inherited actors with Skip_Super are still considered monsters (the Dead* actors); b) if any actor referred to its parent using Super::, its subclassed actor now refers to the the wrong class because Super:: was not changed to the parent's superclass.