Do not copy states from parent actor
Instead of copying all lines from of the state that we want to add points to from the parent, we can just give the player the points and goto the original state in the parent. This produces smaller output files and also fixes problems with ordering the state labels that would arise if a single state of the parent had multiple labels (we didn't handle this possibility before).